Pull out MathJax formatting into a module
